refactor: switch vite demo to typescript
This commit switches the vite demo app to use typescript.
