+++ demo/index.html
@@ -8,8 +8,8 @@
   </head>
   <body>
     <div id="graph-container">
-      <canvas id="mycanvas"></canvas>
+      <canvas id="graphcanvas"></canvas>
     </div>
-    <script type="module" src="/main.js"></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
