+++ demo/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "vite-demo",
-  "version": "0.0.0",
+  "name": "oapi-litegraph-nodegen-demo",
   "private": true,
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "build": "vite build",
     "dev": "vite",
+    "build": "tsc && vite build",
     "preview": "vite preview",
     "sort-package-json": "sort-package-json"
   },
@@ -18,9 +18,10 @@
     "process": "^0.11.10"
   },
   "devDependencies": {
+    "typescript": "~5.6.2",
     "@types/node": "^22.9.0",
     "sort-package-json": "^2.12.0",
-    "vite": "^4.4.0",
+    "vite": "^6.0.1",
     "vite-plugin-node-polyfills": "^0.22.0"
   }
 }
